pandora_box/requirements/base.txt
coverage
Flask==2.0.2
gunicorn
Pillow==9.0.0
pytest
requests